+++ src/main/java/outfoot/outfootserver/friend/service/FriendService.java
@@ -12,6 +12,8 @@ import outfoot.outfootserver.friend.exception.FriendErrorCode;
 import outfoot.outfootserver.friend.exception.FriendException;
 import outfoot.outfootserver.friend.repository.FriendRepository;
 import outfoot.outfootserver.member.domain.Member;
+import outfoot.outfootserver.member.exception.AuthErrorCode;
+import outfoot.outfootserver.member.exception.AuthException;
 import outfoot.outfootserver.member.repository.MemberRepository;
 
 import java.util.List;
@@ -40,7 +42,9 @@ public class FriendService {
     public  void deleteFriend(Member member, Long friendId) { // TODO: 친구 테이블 내 값도 삭제 필요
         Friend friend = friendRepository.findByFromMemberAndId(member, friendId)
                 .orElseThrow(() -> new FriendException(FriendErrorCode.FRIEND_NOT_FOUND));
-
+        if (friend.getFromMember() != member) {
+            throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
+        }
         friendRepository.delete(friend);;
     }
